+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+# 1.0.0-alpha.15
+
+* [FEATURE] add `Query.merge` to allow two queries to be merged
+* [FEATURE] store rich includes to allow for recursively mergable includes
+* [FEATURE] add `associateOnly` as include option that doesn't save the data on included associations, just sets the association
+* [FIX] use transaction in all sequelize calls in `update()`
+* [FIX] coerce `associate` save errors into a `ValidationError` and prefix the property keys with the assoc key
+
 # 1.0.0-alpha.14
 
 * [FIX] when associating belongs-tos in update, check original data for direct id attr, not data from db